project_akhir_pemweb

+++ src/database/migrations/2023_06_21_093910_create_peminjaman_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
     public function up()
     {
         Schema::create('peminjaman', function (Blueprint $table) {
-            $table->integer('id_pinjam')->nullable();
+            $table->bigIncrements('id_pinjam')->nullable();
             $table->integer('id_petugas')->nullable();
             $table->integer('id_anggota')->nullable();
             $table->integer('id_buku')->nullable();
